+++ source/Calamari.Tests/Fixtures/Integration/Scripting/ScriptEngineFixture.cs
@@ -1,11 +1,8 @@
 using System;
 using System.Collections.Generic;
-using Calamari.Common.Features.Processes;
 using Calamari.Common.Features.Scripting;
 using Calamari.Common.Features.Scripts;
-using Calamari.Common.FeatureToggles;
 using Calamari.Common.Plumbing.Logging;
-using Calamari.Common.Plumbing.Variables;
 using Calamari.Testing.Helpers;
 using FluentAssertions;
 using NSubstitute;
@@ -34,18 +31,6 @@ namespace Calamari.Tests.Fixtures.Integration.Scripting
             ScriptSyntax.Bash
         };
 
-        static IEnumerable<TestCaseData> FSharpDeprecationFeatureToggleTestData()
-        {
-            yield return new TestCaseData(new KeyValuePair<string, string>(),
-                                          false);
-            yield return new TestCaseData(
-                                          new KeyValuePair<string, string>(KnownVariables.EnabledFeatureToggles, ""),
-                                          false);
-            yield return new TestCaseData(
-                                          new KeyValuePair<string, string>(KnownVariables.EnabledFeatureToggles, FeatureToggle.FSharpDeprecationFeatureToggle.ToString()),
-                                          true);
-        }
-
         [Test]
         [Category(TestCategory.CompatibleOS.OnlyWindows)]
         public void DeterminesCorrectScriptTypePreferenceOrderWindows()
@@ -66,35 +51,6 @@ namespace Calamari.Tests.Fixtures.Integration.Scripting
             var supportedTypes = engine.GetSupportedTypes();
 
             supportedTypes.Should().Equal(expected);
-        }
-
-        [Test]
-        [TestCaseSource(nameof(FSharpDeprecationFeatureToggleTestData))]
-        public void LogsFSharpDeprecationWarningWhenToggledOn(KeyValuePair<string, string> variableKvp, bool expected)
-        {
-            // Arrange
-            var log = Substitute.For<ILog>();
-            var script = new Script("fakeScript.fsx");
-            var variables = new CalamariVariables
-                { { variableKvp.Key, variableKvp.Value } };
-            var commandResult = new CommandResult("fakeCommand", 0);
-            var commandLineRunner = Substitute.For<ICommandLineRunner>();
-            commandLineRunner.Execute(Arg.Any<CommandLineInvocation>()).Returns(commandResult);
-
-            var engine = new ScriptEngine(new List<IScriptWrapper>(), log);
-
-            // Act
-            engine.Execute(script, variables, commandLineRunner);
-
-            // Assert
-            if (expected)
-            {
-                log.Received(1).Warn(Arg.Is<string>(s => s.StartsWith("Executing FSharp scripts will soon be deprecated")));
-            }
-            else
-            {
-                log.DidNotReceive().Warn(Arg.Any<string>());
-            }
         }
     }
 }
